Renames utils.ts to fieldUtils.ts
