+++ app/form/v1_0/fields/common/fieldUtils.test.ts
@@ -5,7 +5,7 @@ import {
   getMediumLabelClassFor,
   getSmallLabelClassFor,
   validateMaxLength,
-} from './utils'
+} from './fieldUtils'
 import { FieldType } from '../../../../../server/@types/hmpo-form-wizard/enums'
 
 describe('fields/common', () => {
