+++ app/form/v1_0/fields/common/fieldsFactory.ts
@@ -1,6 +1,13 @@
 import FormWizard from 'hmpo-form-wizard'
 import { FieldType, ValidationType } from '../../../../../server/@types/hmpo-form-wizard/enums'
-import { dependentOn, fieldCodeWith, getMediumLabelClassFor, orDivider, validateMaxLength, yesNoOptions } from './utils'
+import {
+  dependentOn,
+  fieldCodeWith,
+  getMediumLabelClassFor,
+  orDivider,
+  validateMaxLength,
+  yesNoOptions,
+} from './fieldUtils'
 import characterLimits from '../../config/characterLimits'
 
 type Section = {
