+++ app/form/v1_0/fields/common/index.ts
@@ -1,4 +1,4 @@
-import * as utils from './utils'
+import * as utils from './fieldUtils'
 import FieldsFactory from './fieldsFactory'
 
 export { utils, FieldsFactory }
