+++ app/form/v1_0/fields/offence-analysis.ts
@@ -2,7 +2,7 @@ import FormWizard from 'hmpo-form-wizard'
 import { FieldsFactory, utils } from './common'
 import { FieldType, ValidationType } from '../../../../server/@types/hmpo-form-wizard/enums'
 import sections from '../config/sections'
-import { dependentOn, yesNoOptions } from './common/utils'
+import { dependentOn, yesNoOptions } from './common/fieldUtils'
 import characterLimits from '../config/characterLimits'
 
 class OffenceAnalysisFieldsFactory extends FieldsFactory {
