+++ app/form/v1_0/fields/personal-relationships-community.ts
@@ -3,7 +3,7 @@ import { FieldType, ValidationType } from '../../../../server/@types/hmpo-form-w
 import { FieldsFactory, utils } from './common'
 import sections from '../config/sections'
 import characterLimits from '../config/characterLimits'
-import { orDivider } from './common/utils'
+import { orDivider } from './common/fieldUtils'
 
 const childrenInformationHint = `
 <p class="govuk-hint">This refers to any children (under 18 years) [subject] has regular contact with, even if they do not have parental responsibility.</p>
